requirements.txt
# Core dependencies for FINMA Career Coach
streamlit>=1.29.0
anthropic>=0.8.0

# Jupyter support
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.27.0

# Additional utilities
python-dotenv>=1.0.0
